refactor : buttion UI --> shared Layer
