+++ src/shared/index.ts
@@ -1,1 +1,2 @@
-export { Pagination } from './ui/pagination'
+export { Pagination } from './ui/Pagination'
+export { Button } from './ui/Button'
